requirements.txt
ai-django-core==5.0.0
appdirs==1.4.4
asgiref==3.3.1
bleach==3.3.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.7
defusedxml==0.7.1
distlib==0.3.1
dj-database-url==0.5.0
dj-rest-auth==1.1.0
Django==3.1.5
django-allauth==0.44.0
django-appconf==1.0.4
django-bootstrap3==14.2.0
django-core==1.4.1
django-cors-headers==3.7.0
django-heroku==0.3.1
django-rest-auth==0.9.5
django-user-accounts==3.0.2
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
filelock==3.0.12
geonode-user-accounts==1.0.15
gunicorn==20.1.0
html2text==2020.1.16
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==2.11.3
MarkupSafe==1.1.1
micropython-fcntl==0.0.3
micropython-ffilib==0.1.2
oauthlib==3.1.0
packaging==20.9
Pillow==8.2.0
pipenv==2020.11.15
psycopg2==2.8.6
psycopg2-binary==2.8.6
pycparser==2.20
PyJWT==2.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
ruamel.yaml==0.17.4
ruamel.yaml.clib==0.2.2
six==1.15.0
sqlparse==0.4.1
uritemplate==3.0.1
urllib3==1.26.4
virtualenv==20.2.2
virtualenv-clone==0.5.4
webencodings==0.5.1
whitenoise==5.2.0
